YUK-15141 Deleting a Device while Selected in Web Commander Causes a Permanent Yukon Exception  - added handling of a paobjectId that no longer exists.

+++ yukon-client/yukon-web/src/com/cannontech/web/bulk/MassDeleteController.java
@@ -27,6 +27,7 @@ import com.cannontech.common.device.groups.editor.dao.DeviceGroupMemberEditorDao
 import com.cannontech.common.device.groups.editor.model.StoredDeviceGroup;
 import com.cannontech.common.device.groups.service.TemporaryDeviceGroupService;
 import com.cannontech.common.device.model.SimpleDevice;
+import com.cannontech.common.pao.dao.PaoPersistenceDao;
 import com.cannontech.common.pao.service.PaoPersistenceService;
 import com.cannontech.common.util.ObjectMapper;
 import com.cannontech.common.util.RecentResultsCache;
@@ -45,6 +46,7 @@ public class MassDeleteController {
     @Autowired private DeviceGroupCollectionHelper deviceGroupCollectionHelper;
     @Autowired private DeviceGroupMemberEditorDao deviceGroupMemberEditorDao;
     @Autowired private PaoLoadingService paoLoadingService;
+    @Autowired private PaoPersistenceDao paoPersistenceDao;
     @Autowired private PaoPersistenceService paoPersistenceService;
     @Autowired private TemporaryDeviceGroupService temporaryDeviceGroupService;    
     
@@ -106,15 +108,11 @@ public class MassDeleteController {
     
     private void processDeviceDelete(SimpleDevice device) {
         try {
-            deviceDao.removeDevice(device);
-        } catch (IllegalArgumentException iae) { // could be thrown by unknown PaoType for DBPersistent (see DeviceFactory)
-
-            try {
+            if (paoPersistenceDao.supports(device)) {
                 paoPersistenceService.deletePao(device);
-            } catch (Exception e) {
-                throw new ProcessingException("Could not delete device: " + paoLoadingService.getDisplayablePao(device).getName() + " (id=" + device.getDeviceId() + ")", e);
+            } else {
+                deviceDao.removeDevice(device);
             }
-
         } catch (DataRetrievalFailureException e) {
             throw new ProcessingException("Could not find device: " + paoLoadingService.getDisplayablePao(device).getName() + " (id=" + device.getDeviceId() + ")", e);
         } catch (Exception e) {
